\name{GeneClusterNet-package}
\alias{GeneClusterNet-package}
\docType{package}
\title{
cluster gene expressions and reconstruct gene regulatory networks
%%\packageTitle{GeneClusterNet}
}
\description{
GeneClusterNet is a contributed R package for reconstructing gene regulatory network from time course gene expression data based on clustering of dynamic gene expressions. It provides functions for gene expression clustering, deciding the optimal number of clusters based on Bayesian Information Criterion (BIC), interpolating expression data for unevenly spaced measurements to have expression data as measured at even time intervals, and applying Dynamic Bayesian Network model to reconstruct gene regulatory networks. It also includes functions for displaying and visualizing clusters and networks. 
}
\details{
\packageDESCRIPTION{GeneClusterNet}
\packageIndices{GeneClusterNet}
}
\author{
\packageAuthor{GeneClusterNet}

Maintainer: \packageMaintainer{GeneClusterNet}
}
\references{
Wang, Y., Xu, M., Wang, Z., Tao, M., Zhu, J., Wang, L., et al. (2012). \emph{How to cluster gene expression dynamics in response to environmental signals}. Briefings in bioinformatics, 13(2), 162-174.

Wang, Y., Berceli, S. A., Garbey, M. and Wu, R. (2016). \emph{Inference of gene regulatory network through adaptive dynamic Beyesian networm modeling}. Technical Report.

R package G1DBN available at \url{https://cran.r-project.org/package=G1DBN}
}
